Updated working model
The modified code for training the model using the datasets as well as the testing code are updated in this commit. The pre-processing code snippets and the accuracy matrix generation snippets are also processed herein. The training datasets are updated as .zip extensions and comprise several datasets (10test, 120test, 120train, 200train, 300train)